Adds check for `response` to allow rendering in ActionMailer
We have a case where we want to render PDFs from an action and include
them attached to emails. ActionMailer::Base does not have a `response`
method so we have to check to see if it's present before trying to call
it. Since it's only there to set the content-disposition in a
ActionController context it seems safe to ignore if it's not there.