Removes prawn plug-in loading
When used within a Rails app and in the context of Bundler, this code
boiled down to only loading the `prawn-table` plug-in. Any other plug-in
that is included in the Gemfile would automatically be loaded by
Bundler before the loading code was called.

For more rationale see my comment here:
https://github.com/cortiz/prawn-rails/pull/29#issuecomment-400425399