Delegate to Prawn::Document for extensions
This fixes loading order issues where Prawn plug-ins are loaded after
prawn-rails and removes the need for the custom loading code.

Let's say you have prawn-rails and prawn-svg in your application. If
prawn-rails was loaded before prawn-svg, then the `svg` extension would
not be available for PrawnRails::Documents. This was caused by
PrawnRails::Documents extending a version of `Prawn::Documents` that did
not have the extension in Prawn::Document.extensions. When a subclass
extends Prawn::Document, it is assigned all of the current extensions
but no new ones are ever added. By delegating to Prawn::Document we can
be sure to always have the full set of extensions as they're applied
when a new Document is instantiated.